Started fleshing out code base with some sensor mamagement code.
sensor.py holds some classes for modeling sensors needed for the higher level code.
web.py is a bare Flask web app that will form the SensorML endpoint.
collector.py holds all the code to pull data out of /sys/ on a regular basis.

None of the above are integrated at this point.